# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is DuckBug.io API.",
        "title": "DuckBug API",
        "termsOfService": "https://duckbug.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@duckbug.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/logs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a log by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Get a log by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/log.Entity"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing log entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Update a log entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Log entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Log update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/log.Update"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated log entry",
                        "schema": {
                            "$ref": "#/definitions/log.Entity"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Log entry not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a log entry by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Delete a log entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Log entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request - when ID is not provided",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - when something goes wrong",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/v1/technologies/{id}": {
            "get": {
                "description": "Get a technology by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "technologies"
                ],
                "summary": "Get a technology by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Technology ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/technology.Entity"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "log.Entity": {
            "type": "object",
            "properties": {
                "context": {
                    "description": "Context can be any JSON value\n@Schema(\n  oneOf={\n    string,\n    object,\n    array,\n    number,\n    boolean,\n    null\n  },\n  example={\"key\":\"value\"}\n)"
                },
                "id": {
                    "type": "string",
                    "example": "a08929b5-d4f0-4ceb-9cfe-bb4fc05b030c"
                },
                "level": {
                    "type": "string",
                    "example": "INFO"
                },
                "message": {
                    "type": "string",
                    "example": "first log message"
                },
                "time": {
                    "type": "integer",
                    "example": 1704067200000
                }
            }
        },
...
        "log.Update": {
            "type": "object",
            "properties": {
                "context": {
                    "description": "Context can be any JSON value\n@Schema(\n  oneOf={\n    string,\n    object,\n    array,\n    number,\n    boolean,\n    null\n  },\n  example={\"key\":\"value\"}\n)"
                },
                "level": {
                    "type": "string",
                    "enum": [
                        "DEBUG",
                        "INFO",
                        "WARN",
                        "ERROR",
                        "FATAL"
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "updated log message"
                }
            }
        },
...
        "technology.Entity": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Mobile app development framework"
                },
                "exampleDsnConnection": {
                    "type": "string",
                    "example": "postgresql://user:password@localhost:5432/mydb"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "React Native"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}